server/requirements.txt
python-dotenv>=1.0.0
supabase>=0.7.0
flask>=2.0.0
flask-cors>=3.0.0